Spring boot app.  User can sign up, log in, edit
account and reset password using spring security (user

data is stored in tables "user", "password", "role" and

"token", which is used for password reset). On the

/fields page fields are created and recorded to the

"field" table (there's also table "option" for options

of "combobox", "checkbox" and "radiobutton" fields) .

They can be modified in modal and deleted. On the main

page any visitor has the opportunity to take a survey,

which consists of active fields from the database. Data

will be sent to server and recorded to the "response"

table, which implements EAV model and stores data in

following format: field_label, response_id, value. At

the same time, data are transferred to the /responses

page using WebSockets.